Added Tucson Marathon.
Although it was in the mid to low forties at the start, I wore my ice
hat.  All the ice was melted by about mile 20 and by then it was in
the sixties, I think.  I ate 11 caffeinated gels in addition to the
double espresso I had when I got up at 4am and the double espresso I
had at 6:30am (for a 7:00am start).