Added today's La Luz run.
Got off to a late start in part due to wanting to let Sophie sleep in,
in part due to wanting to do a little programming.

Standard espresso regimen (double w/oatmeal @ 5:15, another double right
before starting, third double at turnaround).

26F - 42F.  Wore shorts and singlet, Smartwool gloves and ice hat
loaded with ice.  This was the coldest it's been when I've worn my ice
hat, but I figured it wouldn't melt much when it was cold and would be
useful as it warmed up.  I was right.  I'm glad I wore it.

I did not wear microspikes and consequently had to slow down a bunch
toward the top.  I also took a minor spill when I passed some hikers
on the left on my way down.  There was plenty of room to pass, except
there was just enough of a slope on the trail that my poor tread
wasn't enough, so I slipped and fell on my left side.

I'll have to do some further analysis to figure out why I was 25
secs/mile slower than December 30th.  My guess is it was a combination
of trail condition, the later start's leaving me with less initial
caffeine oomph, the later start's requiring me to dodge more hikers,
the added heat *and* me being perhaps not in as good of shape as I was
on the 30th (although I certainly hope I'm not *25* seconds/mile in
worse shape).

FWIW, 2 years ago, when I was in better shape than last year, I
averaged 18:27 min/mile on this run because the conditions were
treacherous.  I actually stopped on the way down to warn some hikers.

In theory a week from today is the Sandia Snowshoe race.  Their
website says its on, but it'll be a 5k, although I saw a text message
that said it was canceled.  Not sure what I'll do if there's no event
then.