Added today's AA pack run.
Did have oatmeal with the hope that it would buffer the ibuprofen, but
still have post run stomach pains.  250 calories per bottle, mostly
finished the second bottle.

I lost 111 due to stoplights.  Temperature was 31F - 37F.  Between it
being cooler than last week and losing slightly less time due to the
traffic, it's not clear that I've improved significantly.

I did ride my bike on Sunday, but I skipped Monday's interval
training.  I did so well on my last Bosque Bike run that I'm hoping to
do well on Saturday and also be able to do a Sunday bike ride and
Monday interval training.  A week from today I'll only be running 6
miles as part of a micro taper in preparation for The Quad and the
Ghost Town 50k, both of which I'll be doing at a reduced effort.