Updated dates for Bosque and AA runs.
I'm going to wait until Monday to do my 24 mile 40# pack run.  I still
have junk in my lungs, but my resting heart rate is back down to 47,
so I'm probably not *that* sick.  Monday is supposed to be colder than
Sunday, so I think I'll finish quicker if I wait until then.  I also
think it's less likely that my illness will come back to get me if I
skip two days instead of one.

I'll then do my "Wednesday" AA run on Thursday so that I have two days
off between the Bosque run and Wednesday, but I normally have three
days off so I don't expect to do particularly well on that run, but to
some extent the hay is in the barn and it's more important to start
the taper fairly healthy than it is to try to wring the last little
bit of training out of my body.