Added today's final pack training run.
I also changed this from a training plan to a training log since all
the hay is in the barn.  I added a prediction as well as a link to a
Ben Skardon video.

This morning was above freezing, but I forgot to record the exact temperature.
I had a double espresso at 5:15 and another about ten minutes before I
ran.  I did not eat breakfast until after.  I wore shorts and a singlet and
my ice hat.

Yesterday evening I went out with Team RWB and had a couple of beers.  I
normally don't drink the night (or day!) before a pack run, but I greatly
appreciate Team RWB and haven't had a chance to get out with them much
this year, so I'm glad I did.