Adds this morning's pack run.  My pace was slow as was my HR.  In addition to this being no extra caffeine and in a semi-fasted state, I have some cold symptoms (the kids got hit first, then Sophie, then I).
The thing that probably slowed me down the most is that I had a 5
pound plate in the top section of my SAR pack and it slipped forward
and would hit me in the head as I ran.  It did not start out too far
forward, but slipped a little with each step for the first quarter
mile or so.  I compensated by bending my neck down, but that shortened
my stride.  It was comically unpleasant.

I also decided to include most of the leg-based events that are
currently on my calendar.  I did not retroactively fill in the various
leg-based events I've done since BMDM (Run as one 5k, Cedro 50k, Elevation
Insanity, Run for the Zoo 1/2 and Armed Forces 1/2).