Added today's pack run.
I'm disappointed in my pace considering my BPM.  I'm not sure if
that's because I ran 13.1 yesterday (albeit at an easy pace) or the
lack of extra caffeine or the heat (although it was raining and that
kept the heat down).  It may be due to my weight loss.

FWIW, my target was 10:18, so I did beat that, but I was working hard
enough toward the end that if it had been a longer run, I think the
wheels would have fallen off.  The only reason it was so short was so
I could link up with Team RWB for the "22 is too many 5k".

OTOH, I wouldn't have been doing a pack run at all, except the Sandias
are closed due to fire concerns.  I think I'm doing OK on the ultra side
of my training while I'm still improving my BMDM base fitness.