Added today's pack run.  I broke 2:30, but @ 132bpm, not 130bpm.  So close!
I was a little slow early on and wound up running at about 150bpm during the
last mile and a half or so, because I really wanted to break 2:30.  Besides,
I knew I was going to spend a bunch of time in a car today and wouldn't get
much more exercise (may miss core).