Added today's pack run.  I did a pack run today as part of my taper for LT100.
I'm pacing Jeff a week from today, so I didn't want to put too much
time on my feet.  I'm still working on losing weight (I had a couple
days of pizza and beer set me back), so I wanted to get the most bang
for my buck, so I ran with my pack again (making today the fourth day
in a row doing so).  I considered running up and down a bit of La Luz,
but decided the chance of me falling and injuring myself was too high,
so my plan was to run 8 miles with an average HR of 150.

I wasn't able to keep my HR that high, probably because this was my
fourth day in a row running with a pack and my body doesn't fully
recover as quickly from that as from running w/o the pack.  However, this
is the fastest I've ever run this distance or greater at *any* HR and
I do not feel like I need an ice bath afterward.

Oh, and I ran 8.56 instead of 8 even because I wanted to go under the
second underpass and up the hill to the other side so I'd get a little
extra uphill and downhill work.  If today weren't a taper day, I'd
probably have done at least 12, maybe 18 miles.