Updated to be more current.  Made a few clarifications.
FWIW, I noticed that my December 8th run is listed as 15 miles "from house",
but that doesn't make sense.  That should be a La Luz run.  However, if I
make that mod, then the rest of my dates are out of phase.  As such, I plan
on redoing the calendar in a separate commit.