Includes today's La Luz pack run.
Today I ran my fastest pace up to the plate and back.  However, I had
to start from the alternate trailhead, which added some distance, so
my time is a little slower than my fastest time on the normal course.

I was not able to get my HR up high and keep it there.  I think that was
due to calorie deficiency.  I deliberately ran a deficit yesterday as
part of my weight loss protocol.  I didn't eat breakfast this morning in
part because I was hurrying to get to a Team RWB after my LL run.  I
didn't take extra calories either, because I thought it was going to
be cool enough that I wouldn't go through both 20 oz. bottles.  HA!

FWIW, I used very warm water in my bottles and regretted it.  It was above
freezing when I started and I also initially didn't like the fact that
I had my heavier gloves on.  Toward the turnaround it was chilly enough
that the heavyness of my gloves wasn't an issue, but of course I was warm
again toward the bottom.

Oh, I also had to stop and pee, so all told I have a lot of room for
imnprovement on this run.