Added today's pack run.
My improvement is a bit misleading in that last week I had a one minute
pause due to a stoplight.  Divided by six, that's ten seconds per mile,
which is how much I "improved" this week.  However, a one minute pause
gives my body a little bit of recovery time, so my speed after that pause
is a little faster.  Additionally, I added half a mile to the distance and
that half mile was on a slope (i.e. half of it was downhill, the other
half uphill), and the time gained by running downhill is never equal to
the time lost running uphill.  So, I did improve over last week, but not
as much as it might appear.

I'm actually happier that I was able to raise my HR by two beats per minute,
although I really was hoping to raise it higher.  Yesterday was taco Tuesday
for Team RWB.  I wound up having a few beers, then eating a bunch of chips
and salsa, then eating a ton of nuts and dates, when I had only planned on
having two or three beers.  I think that left me a little slower than I
would have been had I stayed home.  Next week will be a nice contrast (
although it won't be a completely controlled experiment, since I'm running
a 50k on Saturday).

All in all, I'm very happy with where I am, but I still have plenty of
room for improvement, so I don't want to get complacent.