Includes today's interval pack run.
I did not exercise yesterday.  It was too cold for me to ride my bike
first thing in the morning and it was never particularly sunny, so I
decided to give myself the day off, since I had run pretty hard on
Saturday and I didn't want to let the tiny amount of cold virus left
in me get the upper hand.

So, I had high hopes for today and largely met them.  I'm looking
forward to Wednesday.