Includes today's LL pack run.
Pathetic numbers, but I knew that going in.  Yesterday I did a couple social
runs in addition to my normal Friday 13.1, so all told I "ran" 24 miles in
4:35.  I put ran in quotes because my run with Steve and Ron had some
walking and pausing in it and my run with Margaret was mostly a hike.

I only have four more of these runs left in this training period.  I am
likely to back off my Friday runs for at least a couple of them, just to
see if I can get a better training effect by going in with better rested
legs.  My next one is January 5th, which is only 4 days after Perky's 50k,
so perhaps that might be a good one for me to do less than 13.1 on the
day before.