Includes today's interval pack run.
It was cold (13.5F - 28.4F enough that I wore tights and a long sleeve
shirt under my Team RWB T-shirt.  I had to futz around with my fingers
to keep them warm, which is a hassle in gloves with handhelds.

I haven't yet analyzed my intervals, but they're probably not as good
as last week's, because tights seem to really slow me down.  I don't
know how much of it is psychological and how much is just the extra
heat, but seeing how I have a 50k tomorrow and seeing how I crushed my
Saturday workout, I don't feel too bad about skimping today.  Heck,
not falling, considering all the ice, was pretty important.

FWIW, I spent 834:35 doing activities with my HRM on in 2018. 685:57
of that was running (3,371 miles).  Most of the rest was hiking and
biking.  That means I averaged two hours and seventeen minutes a day
of exercise (with an average HR of 127bpm), not counting the core
workout I do three times a week.

Oh yeah, I managed to get 396,242 ft. of ascent in.  Yay!