Includes yesterday's 50k attempt.
It was cold, snowing and very windy, but the thing that really did me
in was a combination of it being a new course and the snow drifts.

I wound up drinking my 3rd quad espresso before the race started,
rather than waiting until I had finished the southern half, which was
my original plan.  That was a mistake, since I later got mild stomach
cramps.  I also had a lightweight jacket in my pack that I should have
put on, but by the time I realized I wanted it, I had already had
enough experience with cold hands (from having to take a leak,
possibly due to the extra caffeine) that I decided to tough it out,
knowing that I was going to drop at the return to the Indian School
aid station.

The course was well marked and I had it on my GPS watch, but I don't
yet know how to zoom in my GPS watch, so I can only tell when I'm very
far off course, rather than able to use it to tell which of two paths
to choose when there was a fork.

FWIW, nobody even attempted the northern half and only RS even ran the
*entire* southern half.  Two other people quit at the same aid station
that I quit at.  One quit due to being pressed for time, the other quit
at my insistence, because he's relatively new to ultrarunning and I
did not think it would be good for him to be out on his own.

It was still a fun and great day.  Albuquerque doesn't get many
snowstorms like that.