Added today's LL "run".
There was so much snow that I didn't even attempt to break trail on
"new" La Luz.  Instead, I followed other people's footprints up old La
Luz.  As such, I actually made it to where the plate is, although I
couldn't see the plate (it was probably buried) and I would have had
to backed down "new" La Luz to get to it anyway, since old LL comes
out closer to the rock-slide warning sign than the metal plate.

There's so much snow up there that I had to be extremely cautious.
Trees and bushes that normally arc above one's head on the trail are
now eye level due to the height of the snow between one's feet and the
earth.  Since I train with a large external frame pack, that meant I
had to be aware that the pack could catch or I could be pushed,
semi-forcefully, off the path, which could mean a pretty big post-hole
or worse.

As it was, just the added weight of me going downhill had me
postholing in several places where I hadn't, less than an hour ago.

I had already decided to simply enjoy being out in such a beautiful
setting and not to worry about whether I was getting a "proper"
workout.  It would have been nice to take pictures, but as it is I had
to stop to take a leak a couple times and both times my hands got a
bit upset due to the exposure.  I was wearing fairly heavy gloves,
with a "mega" warmer in each (REI was out of the normal sized hand
warmers when I went to restock a while ago).