Includes today's interval pack run.
My intervals were the best ever, although my average pace for all twelve
was 2.1 seconds *slower* than my best, but that's because I got confused
on my sixth interval and cut it short by 15 seconds.

My total time for 13.1 miles doing intervals w/40# pack is my best,
although my HR is higher than normal.  That is both because I was
pushing myself and also due to the heat.  It was 40.0F when I started
and 43.9F when I finished.

Completely through luck, I managed to watch one of the most beautiful
moonsets that I've ever seen.  The full moon was just on the horizon as
I started downhill on Academy during my first loop.  It slowly set as
I slowly jogged down.  It was sublime.