Added today's pack LL run.
Conditions were "better" than my last LL run.  In fact, people had traveled
on the new LL trail, so, unlike last time, I wasn't forced to take any of
the old LL trail.  However, the snow was old and varied greatly in depth.
The metal plate, for example, was buried, yet there were still parts of the
trail that were exposed dirt.  Because the snow was old, I had to dodge a
lot of craters, where people had stepped just a few inches off to the side
of the optimal place to step and had sunk several inches.  Judging from
the holes that these mistakes left, some of them were probably momentarily
frightening for the participant.

So, although I was able to make it to the rock-slide and back w/o having
to take any of the old trail, I was still walking much of the upper portion
and even a fair amount of the lower portion.  I never fell, slid or rolled
my ankle.  That was good enough for me.  I did wind up having to stop to
pee twice, which was disappointing, because I really want to avoid that
on race day.  My hope is that the reason I had to do this today and not
last Saturday (on my 24 mile Bosque Bike pack run) is due to the temperature
being low enough that I wasn't sweating much.

My car thermometer said it was 35F at the start and 38F at the end.  My
guess is it was below freezing further up the mountain.  I used hand warmers
inside my thin (and torn) lightweight REI gloves, at least until a little
past the turnaround.  As I was coming down, some branches tore my ear bud
cable from my iPhone, so i decided to stash my earbuds in my little green
waist belt and since I had to stop to do that, I also took off my gloves.