Includes today's LL pack run.
I certainly ran faster on "this" run in 2016.  However, that used La
Luz, rather than the alternate (and the alternate starts and finishes
with a steep twisty hill), and I only ran 4.51 miles rather than 5.02,
but the big difference was my average HR.  Back then it was 157.
Today it was 146.  So, although my 2016 pace was 12:53 min/mi and
today's was only 13:05, I'm definitely more fit than 2016.

Am I enough more fit to break 4:30?  I think if the weather is
comparable (or better), the answer is yes, but I'm definitely nervous
about how poor my La Luz runs have been this year.  I think running
them slowly and avoiding falls (or worse--some of the snow was
downright treacherous) was the right way to go, but there's no denying
that the seven mile uphill at BMDM from mile 6.5 to mile 13.5 is
taxing.

Oh well, the hay is in the barn, so they say.