Added today's 1/2 distance speed run.
Yesterday was Albuquerque's Team RWB Taco Tuesday one year anniversary
That resulted in me having extra calories last night, so I skipped
breakfast this morning.  I also didn't sleep as well last night as I
might have otherwise.

My stats for today are pretty good.  I took a 33 second hit at the
stoplight in the beginning.  My first mile today was 7:28.2, only the
second time in my life (I think) I've been under 7:30 with a pack, but
the other time was last week and I was 7:27.7 *and* I had an
additional 2 pounds in my pack.

I love my Team RWB buddies, but I am curious as to what I would have
run today with proper fueling and rest yesterday and this morning...

Saturday is King of the Hill.  It's the first year I'll be doing it
and it's also the first year they have a ruck division.  I'm a little
nervous due to being unfamiliar with the course.  I haven't fallen in
several weeks and I'd like to keep it that way.