Includes today's 5k ruck race results.
It was chip timed, I think by CCR Timing, so there's a chance that official
results will eventually be published, but I didn't see any mention of it on
CCR Timing's site and they do list future events, so who knows.

This was their fourth 5k ruck, but the number of ruckers showing up appears
to be fewer and fewer.  I used George's pack and poured a bunch of ice
water on my head from my ice chest and even another water bottle of water
over my head before starting.  I had a double espresso at 5:15am when I
woke up and a quad espresso at 7:55 (it was an 8:30 start).  I also had
200 mg ibuprofen at 7:30.

Since the distance was short, the lack of a hip belt and chest strap was
not an issue.