Added today's unplanned ruck walk.  I had George's ruck in the truck, so ...
I got up and got out the door so I could get to the Alameda Open Space early
enough to do some biking @ 130bpm before starting a Team RWB walk at 8.  I
had George's pack in my truck, but I've already removed the 2.5 pound plate
that I had added to it to top it up past 35 pounds for yesterday's race.
Since George wasn't there, I decided to borrow it one more time, so I put
it on for our hour walk.

Jerry and George wanted to do a small amount of running during the
walk, so I sped up a little for that, but it was basically just a
leisurely walk south on the west side of the Rio Grande (we walked
over the bridge to get to the west side).

On the drive back I cranked the truck heater to 90F (as high as it
goes) and put on the seat warmers to get a little heat training.

Whee!