Added today's Team RWB ruck run as well as yesterday's Chimney-Melon 50k
CM50k was not a race this year, due to the Dragon's Back Ultra being
set for the same day.  I did not want to compete with DBU by offering
a free race on the same day, however, I didn't really want to change
the date that I had set aside for CM50k, so once again CM50k was a
training run, but I did finish about an hour quicker than I did last
year.

I didn't try to get my bike out of the shop on Friday, so I couldn't
pick it up yesterday which means I had no bike to ride this morning.
So, I went to Team RWB's Sunday Fun Day with my ruck and when Nora
wanted to go out for thirty minutes an back for thirty minutes, I
decided that was fine.  I wasn't setting the pace, per-se, and I was
also a little tired from my poor sleep due to all the caffeine (and to
a lesser extent the celebratory beer) yesterday, so that's why my
average HR was only 125 instead of my typical 130.