Includes today's SAR pack run w/Team RWB as well as SAR eval time on my feet.
Today I had two Cibola Search and Rescue evaluations to take:
searching and navigating.  I went to Team RWB's Sunday Runday and when
I saw that George was going to ruck, I put my CSAR pack on and ran
with him.  Later I drove over to the foothills and put my pack on and
kept it on for about three hours.  During that time I had my Ambit on
for some of my walking, but not enough to collect good stats.  I
didn't sit down at all, although I was stopped for much of the time.

I have not yet returned my Rucker 20L for repair.  I need to do that
soon, but when I get it back, I will experiment with running without
the hip belt, because I have not yet been able to find an adjustment
of the straps that works with the hip belt that doesn't beat me up.