Includes today's speed run
I had to do it a little earlier this morning due to an appointment across
town.  As such it was dark when I started.  This resulted in me having a
false start due to leaving my water bottles in the (dark) kitchen.  I
also had to run a little slower to avoid tripping.  On top of that, it
was 17.1F when I started and 18.9F when I finished and I'm out of hand
warmers, so I had to futz around with my hands a lot while running.

So, 21 seconds per mile slower than last year w/2bpm HR higher means
I'm still significantly less fit, but I *appear* to be catching up.
In two days the Black Friday Turkey Trots will give me more data,
since it'll be the same time as last year, although the weather is
likely to be different.