Updated King of the Hill date, also change future La Luz runs to 9.5
So, according to their web site, the King of the Hill race will be
eight days before BMDM, like it was last year.  I've adjusted my
calendar to reflect this.  While doing it I noticed that I have
various oddball distances penned in for my La Luz pack runs, none of
which are particularly likely.

Now that the new parking lot is done, if I make it to the plate and
back it will be pretty close to 9.5 miles.  There may be days when I
have to run less due to snow, although my inclination is to just run a
little more where there isn't snow when I have to turn back early.