Includes today's LL pack run
Many things contributed to today being very slow.  The biggest was
stopping a couple of times to discuss trail conditions with hikers
although the trail condition itself also slowed me down.  There was a
lot of blow-down, including one huge tree.  When I encountered the
tree on the way up I spent a little time looking for a way around it,
but decided I had to slowly go through some of the branches.  On the
way back I didn't look for an alternate way through but it still
slowed me down.

There was enough snow to slow me down, too.  The path was fairly well
worn, but the snow had hardened sufficiently that footprints made for
a fairly uneven surface and I really didn't want to trip or twist an
ankle.

Other contributing factors include it being on a Friday rather than
Saturday (I was up late last night), accidentally taking my hand-held
that lacks a tensioner and not remembering that my phone shuffles
playlists if songs have been shuffled (my iPod *doesn't* do this).