Includes today's pack interval run.
FWIW, today was an experiment that involved less than usual sleep last
night and no calories before running and starting late.  So I went
seventeen and a quarter hours between calories (not counting the
calories in this morning's espressos) and then ran 4.8 miles with
a 40# pack before doing my intervals.  I did OK, considering, but was
too whipped to do the extra 4.2 miles to get up to an even 13.1.