Includes today's pack exercise.  A clip broke early on my descent.
The top clip on the right side shoulder strap broke on my descent.  I
mostly walked from there back to my truck.  I heard a noise as I was
running down and thought I felt a slight change in tension, so I
assumed that the strap had just slipped a little.  However, in general
my straps on my Osprey pack don't slip and it wasn't a noise I heard
before.  I had just put it out of my mind when *POP*, the clip broke
and then I had nothing holding the shoulder strap in place.

Initially I wasn't smart enough to use my right hand to secure the
shoulder strap.  Instead I leaned forward and just expected the pack
to pull me to my right, which it did ... into cactus.  I was pretty
lucky, only a few cactus spines went into my right hand, but after a
couple miles of muscling my way down the hill my right knee started
complaining.  It was then that I realized I could become much more
(but not completely!) symmetric by holding onto the shoulder strap
with my right hand.

Even if that hadn't happened, today's run got off on the wrong foot
due to me not properly tightening the Hammer Gel bottle that I use to
hold my turnaround quad espresso.  It was leaking at the beginning of
my "run", but at the time I didn't even know what the source of the
liquid was.  In fact, I assumed it was coming from the hose from my
bladder, because the drips were from where that was hitting my waist
belt.  Nope.  In 20F weather, messing around with a leak is a PITA.
Even after I tightened the cap, liquid was still coming out from the
hip belt, probably because it had pooled up inside the zippered pocket
that contained the flask, but to be sure I drank the remaining
espresso (essentially a double) in case the flask itself had cracked.
That was too much caffeine and pushed me into the "Oh well, at least
it's a beautiful day" zone.

FWIW, I've had my Osprey pack for several years and have used it for
quite a few search and rescue missions.  A few years ago I trained for
BMDM with it and competed with it, but since it's an internal frame
pack, it was too hot.  I've only been training with it this year as a
stop-gap.  I broke down and ordered a new Tioga a few days ago.  It
arrives on Wednesday.