Includes today's interval run.
Although my entire 13.1 was much slower than last week, it was about
twelve degrees hotter and I had an average HR 5bpm slower today than
last week.

Because my Tioga is doing double duty as my SAR pack, I had to
rearrange it yesterday evening for a SAR mission (which was over well
before I made it to incident base).  I am not particularly happy with
the way it's configured right now, but it works OK for both BMDM
training and actual SAR missions, so it could be much worse.