Includes today's interval run.  I skipped the last 4.22 miles due to rain
It was light rain and was refreshing.  I would have liked to stay out
in it and finish up my run, but previous experience shows that rain
contributes to premature clevis pin failure.  Additionally, my back
still has a hole in it where the pack rubs, so not only did I skip
some mileage, I actually walked a bit after I was done with my
intervals.

2020 is turning out to be a "be satisfied with regrouping" year.