Includes today's < 1/2 distance 40# speed run
Today's run should have been 6.55, but I got distracted yesterday evening
and this morning and wound up doing what I did on Monday, which was also
short.  Oops.

Hard to compare against last year since this year was shorter, but
*might* have had less up and down.  Last year also had me stopped
longer at traffic intersections.  However, it was also significantly
warmer this year.  So, the rough number of 22 seconds/mile slower
is certainly in the ballpark.