Updated my pack weight.  Got curious, so I weighed it.
I was pretty beat when I got home earlier today, so I took things out
of my pack without weighing it.  After I got some rest, I got curious
so I put the things back in (wallet, keys, phone, bladder with
remaining Tailwind) and weighed it.  It was 42.4 and that was after
having consumed about 18 oz. of Tailwind.  So my starting weight was
probably around 43.5 pounds which was not my plan, but at least it is
a lesson in getting prepared earlier than I think I need to.