Includes today's 13.1 mi, 41.3# pack run (w/intervals)
When I weighed my pack this morning, it appeared to be under 40#, so I added
a couple of dense objects, then it weighed 41.3.  I was in a hurry to get out
the door, so that's what I ran with.

FWIW, I don't really trust the electronic scale I'm using.  I miss having a
balance to use for a second opinion.