Includes today's pack run interval training.
My overall minutes per mile is much slower than February 17th, but my
interval pace was significantly faster.  FWIW, although it was warmer
today, it wasn't too much warmer, but today I got hammered both by a
long traffic light as well as having to stand for about a minute as I
waited for the seconds to tick up to the top of the minute.
Additionally, on February 17th I forgot to run the extra .95 miles to
bring my distance up to 6.55 (quarter marathon) from 5.6 (out to AA,
one lap, and back.

So, the faster intervals are a good sign, however the reason I only do
6 today is due to me being on my taper, whereas the reason for only 6
on February 17th is due to recovery from The Quad.