Includes today's pack "run" (hike up old La Luz, run down).
I chose to take my pack off at the top and transfer 24 oz. of water from
a bottle in my pack to my hand-helds, so I was light on the way down.  It
was hot and I wore my ice hat.

My right foot is not fully recovered, so I was pretty nervous doing
this, but so far--and I may change my mind tomorrow--it looks like it
was the right thing to do.

I'm coming off a lot of inactivity due to babying my right foot and
I've lost a ton of fitness and gained a bit of weight, so my last four
runs were all done fasted, which gave me *incredibly* bad stats, but I
did lose some weight.