Includes today's 9.5 mile speed run.
This is my fastest 9.5 mile speed run, but probably only because I
skipped the 9.5 mile distance in my fastest year, due to that
Wednesday being the second of January, and hence I had run the
Foothills Fat Ass 50k the day before.  However, today is only the
sixth, so it's still been less than a week since I ran this year's
Foothills Fat Ass 50k and that also caused me to do my 21 mile Bosque
Bike Path on Sunday with Interval training on Monday, so I'm still, in
essence, recovering from the first.

On top of the above, I've had minor cold symptoms the last couple of
days (and Sophie's been hit harder).