Includes today's 40.4 pack speed run.
I avoided all stoplights, but did have trouble when the ancient caddy
for my Ultimate Direction waterbottle let the bottle fly out to the
ground, so I had to stop and grab the bottle and get it back into the
caddy.

Two years ago I ran this distance at an 8:35 min/mile pace, *but* that
was my best pace for a "full distance" (i.e., not taper) speed ruck
run of that training cycle, so being down seven seconds per mile isn't
as bad as it looks.