Updates one of my La Luz runs from the 2019 season (December 22nd, 2018).
The parking lot was torn up due to construction back then and so I had
to use the alternate trailhead.  That added some extra distance that
my GPS watch had recorded, but when I entered my stats I didn't update
the 9.5 estimate that I had filled in at the beginning of the training
season.