Includes today's reduced distance 40.1# pack run.
Out of an abundance of caution, I chose to run out to Albuquerque Academy,
run one lap and come back.  I did this on 200mg of ibuprofen and a total of
three double espressos (one when I got up, two more a bit before I ran).

I got a late start due to some electrical problems at our house.  It
was fairly hot, so I wore my ice hat.

I haven't been able to uplodad from my watch to Movescount yesterday
or today.  It may be that Suunto has finally pulled the plug on that
functionality. They said they'd be doing it in 2021.  OTOH, I haven't
tried rebooting my laptop yet or submitting from a different computer.
It seems a little strange that there wouldn't be better error messages
or a note on the Movescount page if they really deliberately chose to
end the service.