Includes recent activity, up through today's pack "speed" run.
In addition to having sustained a leg injury, Suunto finally stopped
allowing my watch to upload its info to Movescount.  So ... I wrote
software that converts the file that gets created on my laptop into a
format that I can upload to Strava, but it took me a few days to write
the software.

FWIW, I did OK at the Bosque Bigfoot 50k, but it (or perhaps the
consequent celebratory drinking) may have taken more out of me than I
thought.  My performance Monday and today was abysmal and yesterday I
didn't run at all, although weather contributed to me staying in
yesterday.

I am not sure that doing the entire 24 miles on Saturday is the best
way to proceed.  It may make more sense to just take a slow fun day on
the mountain.  It'd have to be slow because we've gotten a bunch of
snow.