Includes today's speed run.
I wore my ice hat and I had extra Tailwind in a bladder with neither
the bladder nor the Tailwind water counting in my 40.2# pack weight.
I chose not to take an ibuprofen and suspect that's the cause of my
slow time.

OTOH, this is the first time I've done a 13.50 mile speed run, so it's
a PR.  Heh.