Includes today's "race" 1/4 distance pack run.
In the last two years my 8 days out run was the King of the Hill 10k
ruck race down in Los Lunas NM.  This year they're not holding it due
to covid.  I chose to run 6 miles instead of doing 1/4 distance on La
Luz, mostly because I have injured myself on that too many times,
*and* I'm not facing Mineral Hill eight days from now.

When I run KoTH, my pack is 35 pounds, but there's a lot more
elevation gain and loss (about 740 ft. of each) and they start at 8am,
so it's hotter.  As such, I can't normalize my times sufficiently
accurately to say for sure that I'm in better shape than I was last
year (but not as good as shape as two years ago), but that's still my
belief.

I haven't yet decided whether to wear my pack in two days for my 1/4
distance interval training, but I'm leaning toward doing so.