Includes today's Wednesday mountain play, since I had a 33# pack on.
My stats look better than last week, but they're probably not that
much better, because last week I stopped to chat both on the way up
and on the way down.  Today I didn't stop to chat at all, although I
did spend a few minutes moving a tree off the trail.  It was too heavy
to lift, but not too heavy to manipulate.