Includes today's pack speed run.  It was very slow for many reasons.
Probably the biggest is that I was drinking gin for a few hours
yesterday afternoon and early evening due to Joao being in town.
Additionally it was relatively warm (43F) and very slightly rainy.  I
ran in the dark and was also wearing my plushy shoes which have a
messed up lacing system and the laces where bouncing in a way that
kind of freaked me out.  I fell hard yesterday and the laces knocking
on my feet would invoke the sense of tripping.

So, today was largely a pack-run failure, but it's convinced me that I
can't use the plushy shoes for my serious training.  I also need to
review the years where I did better and go back to most of the
practices I was doing then (the big difference being that I will
continue to be fasted at least for a while).  Luckily, it's early
in the training season, so some failed experiments now are OK.

Oh and I still haven't kitted up my Tioga, so once again I was using
my Osprey.