Includes today's speed run.
I was not feeling "right" at the beginning but coaxed myself into
going by allowing myself to run slowly and feel good about just
completing it.

I ran with my ice hat, but it still got warm.  I had a hard time
drawing from my 70 oz. bladder.  It was on top and not in the bladder
compartment.  Yesterday evening when I was putting together my stuff,
I mixed the caffeinated Tailwind in my 40 oz. bladder, but when I went
to put it in my pack I saw there was a pinhole leak at the top, which
is why I then transferred the mixture to my 70 oz. bladder and then
put the bladder on top rather than the bladder compartment.

So, although my initial plan was to run a 10:00 min/mile average pace
for the first half and then run a negative split, I changed my mind
when I was not able to easily drink from the bladder.  Instead I tried
to run at about the same exertion level or, in theory, harder, but my
HRM was obviously too low some times and too high others.  I didn't
look at my average pace until I finished my run.

I did beat the 10:18 min/mile pace needed for a 4:30 marathon, but
that's not saying much since this was a flat course.  OTOH, I had
covid about a month ago and I lost fitness, so I can't expect to be
where I "should" be.