Added 2023 plan and updated README and 2022 to reflect what happened.
I have not spent much time proofreading the changes I've made here.  I'm sure
README would benefit from further updates.  On the other hand, the official
starting date is still away.  I mostly did this work just now because I was
waiting (unsuccessfully) for an important call and it was something I could
work on a little and then drop immediately when the call came in.  However,
since the call never came in, I got to this good stopping point.