[Chore] updates 2023 to be a little more current.
My training schedule is nailed down pretty tight, although it still
includes Academy of Pain and Valenclimb's day, both of which were bad
for my training last year, although that was due to how I did them,
not merely because I did them.  I don't even know that they'll be put
on in 2023 and if so I don't know if it'll make sense for me to do
them, but for now they're still in my "Future" table.