includes my 21 mile 40.9# ruck run from today.  It was pitiful.
My hotspot got the best of me, however, I was ridiculously slow for
the average HR the Coros app is claiming I had.  That may have to do
with how the optical HR works.  If not, I've either lost a lot of
fitness or something else was screwing things up.