includes my pack "speed" run from this morning.
Three things held me back from what I should have been capable of with
my current level of fitness:

  Even with 400mg of ibuprofen, my hot-spot on my right foot slowed me
  down after around 5 miles.  I suspect it's morton's neuroma.

  I got stuck at a traffic light at the main entrance to Albuquerque Academy.
  It seemed like forever, but was probably somewhere between one minute and
  two minutes.  It's hard to tell using just the Strava and Coros data.  With
  my Suunto and Movescount, I think I would have been able to figure it out.

  It was hot.  I wore my ice hat, but it was 42F at the start and 41F
  at the end.

On top of those things, my current level of fitness is degraded in part due
to me not having exercised as much due to the hot-spot and the attendant
weight gain.

The silver lining to my run is that my hot-spot got little better after I
was stopped at the stop light.  If my hot-spot doesn't improve, I'll
probably do poorly at BMDM, *but* if it doesn't get worse, I assume I'll
be able to do OK in ultras where I don't have a pack and I'm not racing
(e.g. Cocodona 250 and Western States), since the lack of a pack stresses
the hot-spot less (as the Foothills 50k proved) and because stopping and
resting the foot provides some relief.